% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dt_emulation.R
\docType{data}
\name{dt_emulation}
\alias{dt_emulation}
\title{Example emulation dataset}
\format{
A list with 4 elements:
\describe{
\item{pde_para_train}{Training set of PDE parameter values.}
\item{pde_para_test}{Test set of PDE parameter values.}
\item{dt_pde_train}{Training set of PDE-based outputs corresponding to
\code{pde_para_train}.}
\item{dt_pde_test}{Test set of PDE-based outputs corresponding to
\code{pde_para_test}.}
}
}
\source{
Simulated data / package example data
}
\usage{
dt_emulation
}
\description{
A list of example training and testing objects used to illustrate the
emulation workflow in the package.
}
\keyword{datasets}
