% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_FF.R
\name{FF_1step_R_sigma2R}
\alias{FF_1step_R_sigma2R}
\title{one time step for the Forward Filter. Computes the FF parameters given the data at the relevant time step and the relevant parameters from the last time step.}
\usage{
FF_1step_R_sigma2R(Yt, Ft, Gt, Wt, Vt, mt_1, Mt_1, nt_1, Dt_1, Rinv, delta = 1)
}
\arguments{
\item{Yt}{The data matrix for epoch t.}

\item{Ft}{The matrix of covariates F_t.}

\item{Gt}{G_t, the beta transition matrix.}

\item{Wt}{left-covariance matrix of the noise parameter of beta_{t-1} | D_{t-1}}

\item{Vt}{left-covariance matrix of the noise parameter of Y}

\item{mt_1}{mean of beta_{t-1} | D_{t-1}}

\item{Mt_1}{left-covariance matrix of beta_{t-1} | D_{t-1}}

\item{nt_1}{the shape parameter, or degrees of freedom, of the right-covariance matrix Sigma | D_{t-1}}

\item{Dt_1}{the scale matrix of the right-covariance matrix Sigma | D_{t-1}}

\item{delta}{The right-variance matrix discount factor}
}
\value{
The mean and covariance matrices m_t and C_t of one filtering step, the updated inverse-Wishart parameters a_t and B_t for the right-covariance matrix, plus other relevant parameters.
}
\description{
one time step for the Forward Filter. Computes the FF parameters given the data at the relevant time step and the relevant parameters from the last time step.
}
