% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_FF.R
\name{FF_sigma2R}
\alias{FF_sigma2R}
\title{Forward Filter. Computes the FF parameters given the data at the relevant time step and the relevant parameters from the last time step.}
\usage{
FF_sigma2R(Y, F_ls, G_ls, W_ls, V_ls, m0, M0, n0, D0, nT, R, delta = 1)
}
\arguments{
\item{Y}{The data matrix.}

\item{m0}{mean of beta_{t-1} | D_{t-1}}

\item{M0}{left-covariance matrix of beta_{t-1} | D_{t-1}}

\item{n0}{the shape parameter, or degrees of freedom, of the right-covariance matrix Sigma | D_{t-1}}

\item{D0}{the scale matrix of the right-covariance matrix Sigma | D_{t-1}}

\item{delta}{The right-variance matrix discount factor}

\item{Ft}{The matrix of covariates F_t.}

\item{Gt}{G_t, the beta transition matrix.}

\item{Wt}{left-covariance matrix of the noise parameter of beta_{t-1} | D_{t-1}}

\item{Vt}{left-covariance matrix of the noise parameter of Y}
}
\value{
The mean and covariance matrices m_t and C_t of one filtering step, the updated inverse-Wishart parameters a_t and B_t for the right-covariance matrix, plus other relevant parameters.
}
\description{
Forward Filter. Computes the FF parameters given the data at the relevant time step and the relevant parameters from the last time step.
}
