% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_FFBS.R
\name{FFBS_predict_MC}
\alias{FFBS_predict_MC}
\title{prediction using monte carlo method to estimate post mean:st, to get mean of Y_new}
\usage{
FFBS_predict_MC(
  nsam,
  Y,
  res_ffbs,
  F_ls,
  F_new_ls,
  input,
  input_new,
  nT,
  phi_para,
  gp_sigma2 = 1.1,
  gp_tau2 = 10^(-4),
  delta = 1
)
}
\description{
prediction using monte carlo method to estimate post mean:st, to get mean of Y_new
}
