% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_FFBS.R
\name{FFBS_sampling}
\alias{FFBS_sampling}
\title{Draw posterior samples from FFBS output (MNIW model)}
\usage{
FFBS_sampling(nsam, para_ffbs, F_ls, G_ls, nT, delta = 1)
}
\arguments{
\item{nsam}{Integer. Number of posterior samples to draw.}

\item{para_ffbs}{List. Output of \code{\link{FFBS}}, containing elements
\code{ff} and \code{bs}.}

\item{F_ls}{Covariate matrix or list of matrices (see \code{\link{FF}}).}

\item{G_ls}{State transition matrix or list (see \code{\link{FF}}).}

\item{nT}{Integer. Number of time steps.}

\item{delta}{Numeric scalar. Discount factor. Defaults to \code{1}.}
}
\value{
A named list of length \code{nT + 1}. Elements \code{"T1"} through
\code{"T<nT>"} are arrays of dimension \code{c(p, q, nsam)} containing
posterior samples of \eqn{\Theta_t}. The element \code{Sigma} is an array
of dimension \code{c(q, q, nsam)} containing posterior samples of
\eqn{\Sigma}.
}
\description{
Given the filtered and smoothed distributions from \code{\link{FFBS}},
draws \code{nsam} joint posterior samples of the state matrices
\eqn{\Theta_1, \ldots, \Theta_T} and the covariance \eqn{\Sigma} using the
MNIW sampler.
}
\seealso{
\code{\link{FFBS}}, \code{\link{MNIW_sampler}}
}
