% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_FFBS.R
\name{FFBS_sampling_I}
\alias{FFBS_sampling_I}
\title{Draw posterior samples from FFBS output (identity right-covariance)}
\usage{
FFBS_sampling_I(nsam, para_ffbs, F_ls, G_ls, nT, delta = 1)
}
\arguments{
\item{nsam}{Integer. Number of posterior samples.}

\item{para_ffbs}{List. Output of \code{\link{FFBS_I}}.}

\item{F_ls}{Covariate matrix or list (see \code{\link{FF}}).}

\item{G_ls}{State transition matrix or list (see \code{\link{FF}}).}

\item{nT}{Integer. Number of time steps.}

\item{delta}{Numeric scalar. Discount factor. Defaults to \code{1}.}
}
\value{
A named list of length \code{nT}. Each element \code{"T<t>"} is an
array of dimension \code{c(p, q, nsam)} with posterior samples of the
state at time \eqn{t}.
}
\description{
Draws \code{nsam} posterior samples of the state \eqn{\Theta_t} from the
smoothed distributions produced by \code{\link{FFBS_I}}, assuming an
identity right-covariance matrix.
}
\seealso{
\code{\link{FFBS_I}}
}
