% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_FFBS.R
\name{FFBS_sampling_sigma2R}
\alias{FFBS_sampling_sigma2R}
\title{Draw posterior samples from FFBS output (scalar sigma-squared-times-R model)}
\usage{
FFBS_sampling_sigma2R(nsam, para_ffbs, F_ls, G_ls, nT, R, delta = 1)
}
\arguments{
\item{nsam}{Integer. Number of posterior samples.}

\item{para_ffbs}{List. Output of \code{\link{FFBS_sigma2R}}.}

\item{F_ls}{Covariate matrix or list (see \code{\link{FF}}).}

\item{G_ls}{State transition matrix or list (see \code{\link{FF}}).}

\item{nT}{Integer. Number of time steps.}

\item{R}{Numeric matrix. Fixed spatial correlation matrix.}

\item{delta}{Numeric scalar. Discount factor. Defaults to \code{1}.}
}
\value{
A named list of length \code{nT + 1}. Elements \code{"T1"} through
\code{"T<nT>"} are arrays of dimension \code{c(p, q, nsam)} with state
samples. The element \code{Sigma} is a vector of length \code{nsam} with
samples of \eqn{\sigma^2}.
}
\description{
Draws \code{nsam} posterior samples of the state \eqn{\Theta_t} and the
scalar variance \eqn{\sigma^2} using the MNIG sampler, given the FFBS output
under the scalar-sigma model.
}
\seealso{
\code{\link{FFBS_sigma2R}}, \code{\link{MNIG_sampler}}
}
