% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{gen_F_ls_AR1}
\alias{gen_F_ls_AR1}
\title{Build AR(1) covariate list from a list of response matrices}
\usage{
gen_F_ls_AR1(Y, nT)
}
\arguments{
\item{Y}{List of length \code{nT}. Each element is the \eqn{N \times S}
data matrix at time \eqn{t}.}

\item{nT}{Integer. Number of time steps.}
}
\value{
A list of length \code{nT} where element \eqn{t} is the covariate
matrix \eqn{F_t}.
}
\description{
Constructs the time-varying covariate list \eqn{F_t} for a first-order
autoregressive model: \eqn{F_t = Y_{t-1}} (with \eqn{F_1 = Y_1}).
}
\seealso{
\code{\link{gen_F_ls_AR2}}, \code{\link{gen_F_ls_AR1_EP}}
}
