% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{gen_F_ls_AR1_EP}
\alias{gen_F_ls_AR1_EP}
\title{Build AR(1) covariate list for the episode-block model}
\usage{
gen_F_ls_AR1_EP(Y, nT, n_b, ind)
}
\arguments{
\item{Y}{List of length \code{nT}. Each element is the full spatial data
matrix at time \eqn{t}.}

\item{nT}{Integer. Number of time steps.}

\item{n_b}{Integer. Number of spatial blocks.}

\item{ind}{Data frame of block indices as returned by
\code{\link{generate_grid}}.}
}
\value{
A list of length \code{nT * n_b} with block-wise AR(1) covariate
matrices.
}
\description{
Generates a flattened covariate list for the episode-partition (EP) model
with AR(1) lags. The output list is indexed by \eqn{(t-1) \times n_b + j},
where \eqn{j} is the block index and \eqn{t} is the time index.
}
\seealso{
\code{\link{gen_F_ls_AR1}}, \code{\link{gen_F_ls_AR2_EP}}
}
