% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{gen_F_ls_AR2}
\alias{gen_F_ls_AR2}
\title{Build AR(2) covariate list from a list of response matrices}
\usage{
gen_F_ls_AR2(Y, nT)
}
\arguments{
\item{Y}{List of length \code{nT}. Each element is the \eqn{N \times S}
data matrix at time \eqn{t}.}

\item{nT}{Integer. Number of time steps.}
}
\value{
A list of length \code{nT} where element \eqn{t} is the
\eqn{N \times 2S} covariate matrix \eqn{F_t}.
}
\description{
Constructs the time-varying covariate list \eqn{F_t} for a second-order
autoregressive model by column-binding the two most recent lags:
\eqn{F_t = [Y_{t-2}, Y_{t-1}]} (with special handling for \eqn{t = 1, 2}).
}
\seealso{
\code{\link{gen_F_ls_AR1}}, \code{\link{gen_F_ls_AR2_EP}}
}
