% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{gen_F_ls_AR2_EP}
\alias{gen_F_ls_AR2_EP}
\title{Build AR(2) covariate list for the episode-block model}
\usage{
gen_F_ls_AR2_EP(Y, nT, n_b, ind)
}
\arguments{
\item{Y}{List of length \code{nT}. Each element is the full spatial data
matrix at time \eqn{t}.}

\item{nT}{Integer. Number of time steps.}

\item{n_b}{Integer. Number of spatial blocks.}

\item{ind}{Data frame of block indices as returned by
\code{\link{generate_grid}}.}
}
\value{
A list of length \code{nT * n_b} with block-wise AR(2) covariate
matrices (each of width \eqn{2 \times \text{bncol}}).
}
\description{
Generates a flattened covariate list for the episode-partition (EP) model
with AR(2) lags. Both lag-1 and lag-2 column blocks are extracted and
column-bound for each spatial block.
}
\seealso{
\code{\link{gen_F_ls_AR2}}, \code{\link{gen_F_ls_AR1_EP}}
}
