% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{recover_from_EP_MC}
\alias{recover_from_EP_MC}
\title{Recover episode-partitioned posterior samples to original time dimension}
\usage{
recover_from_EP_MC(dat_EP, nT_ori, nT, nsam)
}
\arguments{
\item{dat_EP}{List of length \code{nT}. Each element is an array of
dimension \code{c(p, bncol, nsam)}.}

\item{nT_ori}{Integer. Number of original time steps.}

\item{nT}{Integer. Total number of episode time steps.}

\item{nsam}{Integer. Number of posterior samples.}
}
\value{
A named list of length \code{nT_ori}. Each element is an array of
dimension \code{c(p, bncol * n_block, nsam)}.
}
\description{
Reassembles a flattened episode-partitioned list of posterior sample arrays
back into a list indexed by the original \code{nT_ori} time steps.
}
\seealso{
\code{\link{recover_from_EP_exact}}
}
