% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{rmn_chol_more}
\alias{rmn_chol_more}
\title{Generate multivariate normal}
\usage{
rmn_chol_more(nsam, m, RM, RSigma)
}
\arguments{
\item{nsam}{Number of samples}

\item{m}{Mean matrix}

\item{RM}{Upper-triangular Cholesky factor of row covariance matrix}

\item{RSigma}{Upper-triangular Cholesky factor of column covariance matrix}
}
\value{
A random multivariate normal matrix
}
\description{
Generate multivariate normal
}
