% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDBL-package.R
\docType{package}
\name{spDBL-package}
\alias{spDBL}
\alias{spDBL-package}
\title{spDBL: Dynamic Bayesian Learning for Spatiotemporal Mechanistic Models}
\description{
Provides tools for Bayesian learning of spatiotemporal dynamical mechanistic
models, including methods for parameter estimation, simulation, and inference
using hierarchical and state-space modelling approaches. Core computational
routines (forward filter, backward sampler, matrix-normal sampling) are
implemented in C++ via \pkg{RcppEigen}.
}
\author{
\strong{Maintainer}: Xiang Chen \email{xiangchen@ucla.edu}

Authors:
\itemize{
  \item Sudipto Banerjee \email{sudipto@ucla.edu}
}

}
\keyword{internal}
