% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{gen_ffbs_csv}
\alias{gen_ffbs_csv}
\title{Generate FFBS related data}
\usage{
gen_ffbs_csv(N, S, p, nT, path)
}
\arguments{
\item{N}{Number of rows of Y}

\item{S}{Number of columns of Y}

\item{p}{Number of rows of Theta}

\item{nT}{Number of time stamps}

\item{path}{Path for saving generated data}
}
\value{
A list of .csv files: Y's m0 M0 G0 F0 Sigma RSigma(chol(Sigma)) V0 RVO(chol(V0)) W0
}
\description{
Generate FFBS related data
}
