% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{gen_pd_matrix}
\alias{gen_pd_matrix}
\title{Generate random positive definite matrix}
\usage{
gen_pd_matrix(dim)
}
\arguments{
\item{dim}{Number of rows/columns}
}
\value{
A random pd matrix
}
\description{
Generate random positive definite matrix
}
