% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{gen_ran_matrix}
\alias{gen_ran_matrix}
\title{Generate random matrix}
\usage{
gen_ran_matrix(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows}

\item{col}{Number of columns}
}
\value{
A random matrix
}
\description{
Generate random matrix
}
