% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_generate_grid_Dan.R
\name{generate_grid}
\alias{generate_grid}
\title{Generate the indices for a grid for big data traversal. Indexes the grid by a number from 1 to K, where K is the total number of blocks partitioned using the grid.}
\usage{
generate_grid(
  fnrow,
  fncol,
  bnrow,
  bncol,
  traversal.mode = "rowsnake",
  is.flexible = NULL,
  is.flexible.row = FALSE,
  is.flexible.col = FALSE
)
}
\arguments{
\item{fnrow}{The number of rows in the file being read.}

\item{fncol}{The number of columns in the file being read.}

\item{bnrow}{The number of rows in each block being read.}

\item{bncol}{The number of columns in each block being read.}

\item{traversal.mode}{The way to traverse the grid. "rowsnake" refers to alternating left-right traversal. "lr" refers to left-right traversal throughout. Defaults to "rowsnake".}

\item{is.flexible}{Whether or not to throw an error if bnrow or bncol do not perfectly divide fnrow and fncol respectively. If set, will override both is.flexible.row and is.flexible.col. Defaults to NULL.}

\item{is.flexible.row}{Whether or not to throw an error if bnrow does not perfectly divide fnrow. Defaults to FALSE.}

\item{is.flexible.col}{Whether or not to throw an error if bncol does not perfectly divide fncol. Defaults to FALSE.}
}
\value{
A data.frame with the columns for block number, row upper index, row lower index, column left index, and column right index, in that order.
}
\description{
Generate the indices for a grid for big data traversal. Indexes the grid by a number from 1 to K, where K is the total number of blocks partitioned using the grid.
}
