% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_MNIW_sampler.R
\name{MNIW_sampler}
\alias{MNIW_sampler}
\title{Sample from matrix-normal inverse wishart distribution
Y = XB + E, E~MN(O, H, Sigma)
Sigma~IW(v, S)
B | Sigma~MN(C, Vb, Sigma)}
\usage{
MNIW_sampler(nsam, X, v, S, C, Vb)
}
\arguments{
\item{nsam}{Number of samples}

\item{X}{Covariate matrix}

\item{v}{Degrees of freedom of IW}

\item{S}{Scale matrix of IW}

\item{C}{Mean matrix of B | Sigma}

\item{Vb}{Row covariance matrix of  B | Sigma}

\item{H}{Row covariance matrix of Y}
}
\value{
List of samples of B and Sigma
}
\description{
Sample from matrix-normal inverse wishart distribution
Y = XB + E, E~MN(O, H, Sigma)
Sigma~IW(v, S)
B | Sigma~MN(C, Vb, Sigma)
}
