% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_data.R
\name{rmn_chol_more}
\alias{rmn_chol_more}
\title{Draw multiple samples from a matrix-normal distribution (Cholesky parameterisation)}
\usage{
rmn_chol_more(nsam, m, RM, RSigma)
}
\arguments{
\item{nsam}{Integer. Number of samples to draw.}

\item{m}{Numeric matrix. Mean matrix (\eqn{p \times S}).}

\item{RM}{Numeric matrix. Upper-triangular Cholesky factor of the row
covariance \eqn{U} (\eqn{p \times p}).}

\item{RSigma}{Numeric matrix. Upper-triangular Cholesky factor of the column
covariance \eqn{\Sigma} (\eqn{S \times S}).}
}
\value{
An array of dimension \code{c(p, S, nsam)}.
}
\description{
Draws \code{nsam} independent samples from \eqn{MN(m, U, \Sigma)} using
the Cholesky factors of \eqn{U} and \eqn{\Sigma}.
}
\seealso{
\code{\link{rmn_chol}}
}
